requirements.txt
python-telegram-bot
python-dotenv
yt_dlp
pyktok